Validate draft and revision request submission
