+++ podium-gateway/src/main/java/nl/thehyve/podium/service/RequestService.java
@@ -13,6 +13,7 @@ import nl.thehyve.podium.common.config.FilterValues;
 import nl.thehyve.podium.common.enumeration.*;
 import nl.thehyve.podium.common.exceptions.AccessDenied;
 import nl.thehyve.podium.common.exceptions.ActionNotAllowed;
+import nl.thehyve.podium.common.exceptions.InvalidRequest;
 import nl.thehyve.podium.common.exceptions.ResourceNotFound;
 import nl.thehyve.podium.common.security.AuthenticatedUser;
 import nl.thehyve.podium.common.security.AuthorityConstants;
@@ -32,6 +33,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
+import javax.validation.ConstraintViolation;
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -62,6 +67,8 @@ public class RequestService {
     @Autowired
     private StatusUpdateEventService statusUpdateEventService;
 
+    @Autowired
+    private RequestReviewProcessService requestReviewProcessService;
 
     @PostConstruct
     private void init() {
@@ -105,6 +112,16 @@ public class RequestService {
             }
         }
         return result;
+    }
+
+    public static void validateRequest(RequestRepresentation request) {
+        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        Validator validator = factory.getValidator();
+
+        Set<ConstraintViolation<RequestRepresentation>> requestConstraintViolations = validator.validate(request);
+        if (!requestConstraintViolations.isEmpty()) {
+            throw new InvalidRequest("Invalid request", requestConstraintViolations);
+        }
     }
 
     /**
@@ -219,7 +236,39 @@ public class RequestService {
         return getCountsForOrganisations(organisationUuids);
     }
 
+    /**
+     * Submit the request by uuid.
+     *
+     * @param user the current user, submitting the request
+     * @param uuid the uuid of the request
+     * @return the updated request
+     * @throws ActionNotAllowed if the request is not in status 'Revision'.
+     */
+    public RequestRepresentation submitRevision(AuthenticatedUser user, UUID uuid) throws ActionNotAllowed {
+        Request request = requestRepository.findOneByUuid(uuid);
 
+        log.debug("Access and status checks...");
+        AccessCheckHelper.checkRequester(user, request);
+        AccessCheckHelper.checkReviewStatus(request, RequestReviewStatus.Revision);
+
+        log.debug("Validating request data.");
+        RequestRepresentation requestData = requestMapper.requestToRequestDTO(request);
+        validateRequest(requestData);
+
+        // Update the request details with the updated revision details
+        request.setRequestDetail(request.getRevisionDetail());
+        requestRepository.save(request);
+
+        // Submit the request for validation by the organisation coordinator
+        requestReviewProcessService.submitForValidation(user, request.getRequestReviewProcess());
+
+        request = requestRepository.findOneByUuid(uuid);
+        RequestRepresentation requestRepresentation = requestMapper.extendedRequestToRequestDTO(request);
+
+        statusUpdateEventService.publishReviewStatusUpdate(user, RequestReviewStatus.Revision, request, null);
+
+        return requestRepresentation;
+    }
 
     /**
      *  Get the request
