+++ podium-gateway/src/main/webapp/app/request/form/request-form.component.html
@@ -17,7 +17,7 @@
     <div class="card-container"
         [ngClass]="{
             'col-md-6 offset-md-3': !isInRevision,
-            'col-md-12 nested-request-form': isInRevision
+            'nested-request-form': isInRevision
         }">
 
         <form #requestForm="ngForm" class="form" role="form">
